+++ pmc_automation_tools/driver/common.py
@@ -60,6 +60,13 @@ _wait_until = {
     EXISTS : None
 }
 
+_wait_untils = {
+    VISIBLE: EC.presence_of_all_elements_located,
+    # INVISIBLE : EC.invisibility_of_element_located,
+    # CLICKABLE : EC.element_to_be_clickable,
+    # EXISTS : None
+}
+
 
 class PlexDriver(ABC):
     def __init__(self, environment: Literal['ux', 'classic'], *args, driver_type: Literal['edge', 'chrome']='edge', **kwargs):
@@ -97,6 +104,46 @@ class PlexDriver(ABC):
         self.latest_driver_version_file = os.path.join(self.resource_path, 'driver_version.txt')
 
 
+    def wait_for_elements(self, selector:Union[tuple[str, str], str], *args, driver=None, timeout:int=15, type=VISIBLE, ignore_exception: bool=False, element_class:object=None) -> 'PlexElement':
+        """Wait until an element meets specified criteria.
+
+        Wrapper for Selenium WebDriverWait function for common Plex usage.
+
+        Args:
+            selector (tuple[str, str], str): Selenium style element selector e.g. wait_for_element((By.NAME, 'ElementName')).
+            *args: If sending a str as the selector, expects the value as an additional positional argument. e.g. wait_for_element(By.NAME, 'ElementName').
+            driver (WebDriver|PlexElement, optional): root WebDriver to use for locating the element. Defaults to None.
+            timeout (int, optional): Time to wait until timeout is triggered. Defaults to 15.
+            type (str, optional): Type of wait to be performed. Defaults to VISIBLE.
+            ignore_exception (bool, optional): Don't raise an exception if the timeout is reached. Defaults to False.
+            element_class (object, optional): class of the WebElement to return. Defaults to None.
+
+        Raises:
+            exception: re-raises exception from failing to meet the expected condition
+
+        Returns:
+            PlexElement: PlexElement based on search criteria
+        """
+        if isinstance(selector, tuple) and len(selector) == 2:
+            by, value = selector
+        elif isinstance(selector, str) and len(args) > 0:
+            by = selector
+            value = args[0]
+        else:
+            raise TypeError('selector argument not instance of tuple or did not receive 2 positional arguments for "by" and "value".')
+        try:
+            driver = driver or self.driver
+            element_condition = _wait_untils.get(type)
+            if element_condition:
+                _elements = WebDriverWait(driver, timeout).until(element_condition((by, value)))
+            element_class = element_class or PlexElement
+            return [element_class(_el, self) for _el in _elements] #element_class(driver.find_element(by, value), self)
+
+        except (TimeoutException, StaleElementReferenceException, NoSuchElementException):
+            if ignore_exception:
+                return None
+            raise
+
     def wait_for_element(self, selector:Union[tuple[str, str], str], *args, driver=None, timeout:int=15, type=VISIBLE, ignore_exception: bool=False, element_class:object=None) -> 'PlexElement':
         """Wait until an element meets specified criteria.
 
@@ -128,9 +175,9 @@ class PlexDriver(ABC):
             driver = driver or self.driver
             element_condition = _wait_until.get(type)
             if element_condition:
-                WebDriverWait(driver, timeout).until(element_condition(selector))
+                WebDriverWait(driver, timeout).until(element_condition((by, value)))
             element_class = element_class or PlexElement
-            return element_class(driver.find_element(*selector), self)
+            return element_class(driver.find_element(by, value), self)
 
         except (TimeoutException, StaleElementReferenceException, NoSuchElementException):
             if ignore_exception: